Spliting server tests into multiple files
