+++ server/tests/simple.rs
@@ -1,72 +1,10 @@
 
-//! ===================================================================================
-//! *****************       READ THIS BEFORE ADDING NEW TESTS         *****************
-//! ===================================================================================
-//! The `aaa_start_stop_test` test is responsible for starting and stopping the
-//! server, and the rest of the tests will wait until they confirm the server is alive
-//! ===================================================================================
-
 use std::{time::{Duration, Instant}};
 use tokio::task;
 use reqwest::{Client, Error};
 
-/// Convenience to declare a literal to use like a constant
-macro_rules! decl_lit { ($NAME:ident!() => $LIT:tt) => { macro_rules! $NAME { () => { $LIT }; } }; }
-
-macro_rules! server_url { () => { "http://127.0.0.1:8000" }; }
-
-/// Waits for the server to become available
-async fn wait_for_server() -> Result<(), Error> {
-    //TODO: Do we want a new command that returns the server status, as opposed to a path status?
-    const URL: &str = concat!(
-        server_url!(),
-        "/", "status",
-        "/", "-"
-    );
-    loop {
-        match reqwest::get(URL).await {
-            Ok(response) => {
-                if response.status().is_success() {
-                    return Ok(())
-                }
-            },
-            Err(_) => {}
-        }
-        tokio::time::sleep(Duration::from_millis(5)).await;
-    }
-}
-
-/// Starts the server, and then dispatches a `stop?wait_for_idle` request
-#[tokio::test]
-async fn aaa_start_stop_test() -> Result<(), Error> {
-
-    //Start the server
-    let mut server_proc = std::process::Command::new(env!("CARGO_BIN_EXE_mork_server")).spawn().unwrap();
-    // println!("start_stop_test: Server start initiated");
-
-    //Give the server time to start, and give the other tests a chance to start using the server
-    tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-
-    // println!("start_stop_test: Sending stop request");
-    const STOP_URL: &str = concat!( 
-        server_url!(),
-        "/", "stop",
-        "/?", "wait_for_idle",
-    );
-    let response = reqwest::get(STOP_URL).await?;
-    if response.status().is_success() {
-        // println!("start_stop_test: Received response: {}", response.text().await?);
-    } else {
-        panic!("start_stop_test: Failed to send stop request: {}", response.status());
-    }
-
-    //Wait for the server to exit
-    // println!("start_stop_test: Waiting for server to stop");
-    server_proc.wait().unwrap();
-
-    // println!("start_stop_test: Server stopped");
-    Ok(())
-}
+mod common;
+use common::*;
 
 /// Opens one client request
 #[tokio::test]
@@ -930,166 +868,6 @@ async fn transform_ooga_booga_regression_test() -> Result<(), Error> {
     Ok(())
 }
 
-#[tokio::test]
-async fn metta_thread_basic() -> Result<(), Error> {
-    decl_lit!{def_expr!()           => "(def $loc $step $p $t)"        }
-    decl_lit!{in_expr!()             => "((metta_thread_basic in)  $v)"}
-    decl_lit!{out_expr!()            => "((metta_thread_basic out) $v)"}
-    decl_lit!{id_pattern_template!() => "$v"                           }
-
-    const UPLOAD_URL: &str = concat!(
-        server_url!(),
-        "/", "upload",
-        "/", id_pattern_template!(),
-        "/", in_expr!(),
-    );
-    const UPLOAD_PAYLOAD : &str = concat!(""
-        , "\n(val a b)"
-        , "\n(val c d)"
-        , "\n(val e f)"
-        , "\n(val g h)"
-    );
-    const UPLOAD_STATUS: &str = concat!( 
-        server_url!(),
-        "/", "status",
-        "/", in_expr!(),
-    );
-
-    const DEF_UPLOAD_URL: &str = concat!(
-        server_url!(),
-        "/", "upload",
-        "/", def_expr!(),
-        "/", def_expr!(),
-    );
-    const DEF_UPLOAD_PAYLOAD : &str = concat!(""
-    , "\n(def metta_thread_basic 0"
-    , "\n     (, ((metta_thread_basic in) (val $x $y))"
-    , "\n        (def metta_thread_basic 1 $p $t)"
-    , "\n     )"
-    , "\n     (, (exec ((metta_thread_basic state) \"cleanup\") (, (val $y $x)) (,))"
-    , "\n        (exec (metta_thread_basic \"asap\") $p $t)"
-    , "\n        ((metta_thread_basic out) (val $x $y))"
-    , "\n     )"
-    , "\n)"
-
-    , "\n(def metta_thread_basic 1"
-    , "\n     (, (exec ((metta_thread_basic state) $p) (, (val $u $v)) (,)) "
-    , "\n     )"
-    , "\n     (, ((metta_thread_basic out) (val $u $v))"
-    , "\n     )"
-    , "\n)"
-    );
-    const DEF_UPLOAD_STATUS: &str = concat!( 
-        server_url!(),
-        "/", "status",
-        "/", "(def metta_thread_basic $step $p $t)",
-    );
-
-    const METTA_THREAD_RUNNER : &str = concat!(
-        server_url!(),
-        "/",  "metta_thread",
-        "/" , "(exec ($t_id \"asap\")   (, (def $t_id 0 $p $t) )   (, (exec ($t_id \"asap\") $p $t) )   )",
-        "/?", "location=metta_thread_basic"
-    );
-    const METTA_THREAD_RUNNER_STATUS: &str = concat!( 
-        server_url!(),
-        "/", "status",
-        "/", "(exec metta_thread_basic)",
-    );
-
-
-    const METTA_THREAD_CLEAR_STATE : &str = concat!(
-        server_url!(),
-        "/", "metta_thread",
-        "/", "(exec ($t_id \"asap\") (,) (,))",
-        "/?", "location=(metta_thread_basic state)"
-    );
-    const METTA_THREAD_STATE_STATUS: &str = concat!( 
-        server_url!(),
-        "/", "status",
-        "/", "(exec (metta_thread_basic state))",
-    );
-
-    const EXPORT_URL : &str = concat!(
-        server_url!(),
-        "/", "export",
-        "/", out_expr!(),
-        "/", id_pattern_template!(),
-    );
-
-    wait_for_server().await.unwrap();
-
-    // upload vals
-    let response = reqwest::Client::new().post(UPLOAD_URL).body(UPLOAD_PAYLOAD).send().await.unwrap();
-    if !response.status().is_success() {
-        panic!("Error response: {} - {}", response.status(), response.text().await.unwrap())
-    }
-    println!("Upload response: {}", response.text().await.unwrap());
-
-
-    // defs
-    let response = reqwest::Client::new().post(DEF_UPLOAD_URL).body(DEF_UPLOAD_PAYLOAD).send().await.unwrap();
-    if !response.status().is_success() {
-        panic!("Error response: {} - {}", response.status(), response.text().await.unwrap())
-    }
-    println!("Def Upload response: {}", response.text().await.unwrap());
-
-    wait_for_url_eq_status(UPLOAD_STATUS,      "pathClear").await.unwrap();
-    wait_for_url_eq_status(DEF_UPLOAD_STATUS,  "pathClear").await.unwrap();
-
-    // exec runner
-    let response = reqwest::get(METTA_THREAD_RUNNER).await.unwrap();
-    if !response.status().is_success() {
-        panic!("Error response: {} - {}", response.status(), response.text().await.unwrap())
-    }
-    println!("Exec Runner response: {}", response.text().await.unwrap());
-
-    wait_for_url_eq_status(METTA_THREAD_RUNNER_STATUS, "pathClear").await.unwrap();
-
-
-    // exec state cleanup
-    let response = reqwest::get(METTA_THREAD_CLEAR_STATE).await.unwrap();
-    if !response.status().is_success() {
-        panic!("Error response: {} - {}", response.status(), response.text().await.unwrap())
-    }
-    println!("Exec Runner response: {}", response.text().await.unwrap());
-
-
-    wait_for_url_eq_status(METTA_THREAD_STATE_STATUS, "pathClear").await.unwrap();
-
-
-    tokio::time::sleep(core::time::Duration::from_millis(10)).await;
-    // //////////
-    // EXPORT //
-    // ////////
-
-    let export_response = reqwest::get(EXPORT_URL).await.unwrap();
-    if !export_response.status().is_success() {
-        panic!("Error response: {} - {}", export_response.status(), export_response.text().await.unwrap())
-    }
-    let export_response_text = export_response.text().await.unwrap();
-    println!("Export response:\n{}", export_response_text);
-
-
-    const OUT_LIST : &str = concat!(""
-    , "(val a b)\n"
-    , "(val b a)\n"
-    , "(val c d)\n"
-    , "(val d c)\n"
-    , "(val e f)\n"
-    , "(val f e)\n"
-    , "(val g h)\n"
-    , "(val h g)\n"
-    );
-    core::assert_eq!(
-        OUT_LIST,
-        export_response_text
-    );
-
-    Ok(())
-
-}
-
 /// Tests that it's at least possible to write to the root of the space
 ///
 /// This test is ignored because it pollutes the root of the space with crap and holds a lock that
@@ -1126,33 +904,4 @@ async fn root_writer_test() -> Result<(), Error> {
     }
 
     panic!()
-}
-
-
-async fn wait_for_url_eq_status(url : &str, expected_status : &str) ->Result<(),reqwest::Error> {
-    wait_for_url_status::<true>(url, expected_status).await
-}
-async fn wait_for_url_ne_status(url : &str, expected_status : &str) ->Result<(),reqwest::Error> {
-    wait_for_url_status::<false>(url, expected_status).await
-}
-async fn wait_for_url_status<const EQ : bool>(url : &str, expected_status : &str) ->Result<(),reqwest::Error> {
-    loop {
-        std::thread::sleep(std::time::Duration::from_millis(10));
-        let response = reqwest::get(url).await?;
-        assert!(response.status().is_success());
-        let response_text = response.text().await?;
-        println!("Status (polling) response: {}", response_text);
-        let response_json: serde_json::Value = serde_json::from_str(&response_text).unwrap();
-        if EQ {
-
-            if response_json.get("status").unwrap().as_str().unwrap() == expected_status {
-                break
-            }
-        } else {
-            if response_json.get("status").unwrap().as_str().unwrap() != expected_status {
-                break
-            }
-        }
-    }
-    Ok(())
 }
